Rework subpath extraction from endpoint URL
The extraction for 'end of url' with path variables seems
to behave differently than I expected (ie. in SpringBoot);
so updated the test to reproduce the error and fixed the
extraction by directly accessing the request.

I left the Optional Path Variable path for the sake of
swagger doc generation, which is an hack so please mind while
revieweing.